refactoring: 분류에 맞게 폴더명 변경
